# API_documentation/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/mastertables/employment_status": {
      "get": {
        "operationId": "EmploymentStatusController_getEmployeeStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Here are the list of available employment status"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all employment status",
        "tags": [
          "Mastertables"
        ]
      },
      "post": {
        "operationId": "EmploymentStatusController_createEmployeeStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create employee status.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee status created successfully."
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create new employee status.",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/employment_status/{id}": {
      "get": {
        "operationId": "EmploymentStatusController_getEmployeeStat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Here is the employment status."
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get an employment status.",
        "tags": [
          "Mastertables"
        ]
      },
      "put": {
        "operationId": "EmploymentStatusController_updateEmployeeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmpStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee status details updated successfully."
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Updating employee status details.",
        "tags": [
          "Mastertables"
        ]
      }
    },
...
    "/mastertables/positions": {
      "get": {
        "operationId": "PositionController_getPositions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of positions retrieve"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all positions",
        "tags": [
          "Mastertables"
        ]
      },
      "post": {
        "operationId": "PositionController_createPosition",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create Position",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new position",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/positions/{id}": {
      "get": {
        "operationId": "PositionController_getPosition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Here is the position."
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a position.",
        "tags": [
          "Mastertables"
        ]
      },
      "put": {
        "operationId": "PositionController_updatePosition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Payload to update Position information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a current position information",
        "tags": [
          "Mastertables"
        ]
      }
    },
...
  },
  "info": {
    "title": "ABAS v3 API",
    "description": "API documentation for ABAS v3 project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      },
      "security_clearance": {
        "type": "apiKey",
        "name": "Security-Clearance",
        "in": "header",
        "description": "Required security clearance level (documentational only)."
      }
    },
    "schemas": {
...
      "CreateEmployeeStatusDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateEmpStatusDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePositionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Software Engineer",
            "description": "The name of the position"
          },
          "sorting": {
            "type": "number",
            "example": 2,
            "description": "Sorting number of the position"
          },
          "department": {
            "type": "number",
            "example": "human resources = 1, information technology = 2, accounting = 6",
            "description": "The Department where the position is available"
          },
          "status": {
            "type": "number",
            "example": "active or inactive",
            "description": "active = 1, inactive = 0"
          }
        },
        "required": [
          "name",
          "sorting",
          "department",
          "status"
        ]
      },
      "UpdatePositionDto": {
        "type": "object",
        "properties": {
          "position_name": {
            "type": "string",
            "example": "New Position name",
            "description": "If you want to update the Position Name"
          },
          "sorting": {
            "type": "number",
            "example": 2,
            "description": "Sorting number of the position"
          },
          "department": {
            "type": "number",
            "example": "human resources, information technology, accounting",
            "description": "The Department where the position is available"
          },
          "status": {
            "type": "number",
            "example": "active or inactive",
            "description": "active = 1, inactive = 0"
          }
        },
        "required": [
          "position_name",
          "sorting",
          "department",
          "status"
        ]
      },
...
    }
  }
}